fix: schema

+++ db/migrate/20220829032943_create_weathers.rb
@@ -4,7 +4,7 @@ class CreateWeathers < ActiveRecord::Migration[6.1]
   def change
     create_table :weathers do |t|
       t.datetime :weather_time
-      t.integer :temperature
+      t.float :temperature
       t.timestamps
     end
   end
